requirements.txt
django-heroku
gunicorn
psycopg2-binary
social-auth-app-django
python-decouple